# api/openapi.yaml
openapi: 3.1.0
info:
  title: Samyama Graph Database API
  version: 1.8.0
  description: |
    HTTP API for the Samyama high-performance distributed graph database.
    Supports OpenCypher queries, graph status, and CRUD operations.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/query:
    post:
      operationId: executeQuery
      summary: Execute a Cypher query
      description: |
        Execute an OpenCypher query against the graph database.
        Supports both read (MATCH) and write (CREATE, SET, DELETE, MERGE) queries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            examples:
              matchAll:
                summary: Match all Person nodes
                value:
                  query: "MATCH (n:Person) RETURN n"
              matchWithFilter:
                summary: Match with WHERE filter
                value:
                  query: "MATCH (n:Person) WHERE n.age > 25 RETURN n.name, n.age"
              createNode:
                summary: Create a new node
                value:
                  query: 'CREATE (n:Person {name: "Alice", age: 30})'
              traversal:
                summary: Traverse relationships
                value:
                  query: "MATCH (a:Person)-[:KNOWS]->(b:Person) RETURN a.name, b.name"
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Query error (parse error, execution error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/begin:
    post:
      operationId: beginTransaction
      summary: Begin a transaction
      description: |
        Opens a transaction and returns its id. Send statements to `/api/query`
        with `"tx": <id>`, then commit or roll back.

        The transaction holds the graph's write lock until it ends: no other
        request can read or write while it is open. It is rolled back
        automatically after `timeout_seconds` (`SAMYAMA_TX_TIMEOUT_SECS`,
        default 30). Isolation is serializable in effect, since nothing runs
        concurrently with an open transaction.
      responses:
        '200':
          description: Transaction opened
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  version:
                    type: integer
                  timeout_seconds:
                    type: integer
        '409':
          description: A transaction could not be opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/{id}/commit:
    post:
      operationId: commitTransaction
      summary: Commit a transaction
      description: Keeps every write the transaction made, persists them, and releases the lock.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Committed
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  committed:
                    type: boolean
                  version:
                    type: integer
        '404':
          description: No such open transaction (already ended, or timed out and rolled back)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The transaction timed out and was rolled back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/{id}/rollback:
    post:
      operationId: rollbackTransaction
      summary: Roll back a transaction
      description: |
        Undoes everything the transaction did -- property and label writes,
        created nodes and relationships, and deletions, which are restored with
        their properties and relationships -- and releases the lock.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rolled back
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  rolled_back:
                    type: boolean
        '404':
          description: No such open transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: An OpenCypher query string
          examples:
            - "MATCH (n:Person) RETURN n"
        graph:
          type: string
          description: "Target graph/tenant name (default: 'default')"
          default: default
        cache:
          type: boolean
          description: >-
            Serve this query from the result cache when the graph has not
            changed since the answer was computed. Omitted means the server
            default, which is off unless SAMYAMA_RESULT_CACHE=1. Send false to
            bypass a warm cache for a latency measurement.
        tx:
          type: string
          description: >-
            Run the statement inside this open transaction, from
            `POST /api/tx/begin`. Its writes are visible only inside the
            transaction until `POST /api/tx/{id}/commit`, and
            `POST /api/tx/{id}/rollback` undoes them.
    QueryResponse:
      type: object
      properties:
        nodes:
          type: array
          description: Graph nodes referenced in the result (for visualization)
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          description: Graph edges referenced in the result (for visualization)
          items:
            $ref: '#/components/schemas/GraphEdge'
        columns:
          type: array
          description: Column names for the tabular result
          items:
            type: string
        records:
          type: array
          description: Tabular result rows, one array per record
          items:
            type: array
            items: {}
        engine_version:
          type: string
          description: The build that produced this result (TRUST-06)
        snapshot_version:
          type: integer
          format: int64
          description: >-
            The MVCC version the read saw, tying the result to the data that
            produced it. Identifies the snapshot; it is not a content hash.
        cached:
          type: boolean
          description: >-
            Whether these rows were served from the result cache rather than
            computed. A latency measured over a hit is not engine latency.
        notifications:
          type: array
          description: >-
            Diagnostics the statement produced. Empty for almost every query.
            A notification reports that the engine made a semantic choice the
            query did not state and that the choice was observable in the rows.
          items:
            $ref: '#/components/schemas/Notification'
    Notification:
      type: object
      properties:
        code:
          type: string
          description: Stable identifier a client can match on
          examples:
            - "Samyama.Notification.Statement.PathModeAffectsResult"
        severity:
          type: string
          examples:
            - INFORMATION
        title:
          type: string
        description:
          type: string
    GraphNode:
      type: object
      properties:
        id:
          type: string
          description: Node ID
        labels:
          type: array
          items:
            type: string
          description: Node labels
        properties:
          type: object
          additionalProperties: true
          description: Node properties
    GraphEdge:
      type: object
      properties:
        id:
          type: string
          description: Edge ID
        source:
          type: string
          description: Source node ID
        target:
          type: string
          description: Target node ID
        type:
          type: string
          description: Relationship type
        properties:
          type: object
          additionalProperties: true
          description: Edge properties
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...